🚚♻️🏷️ Start splitting tests. Extract shared to utils + MultiType
